Split release artifacts assertion from get_distros script
... to run get_distros without any venv or container prerequesit

CMK-17439
