+++ buildscripts/scripts/assert-release-build-artifacts.groovy
@@ -20,7 +20,7 @@ def main() {
                 withEnv(["PYTHONUNBUFFERED=1"]) {
                     dir("${checkout_dir}") {
                         sh(script: """scripts/run-pipenv run \
-                        buildscripts/scripts/get_distros.py \
+                        buildscripts/scripts/assert_build_artifacts.py \
                         --editions_file "${checkout_dir}/editions.yml" \
                         assert_build_artifacts \
                         --version "${VERSION_TO_CHECK}" \
