+++ buildscripts/scripts/utils/versioning.groovy
@@ -101,43 +101,37 @@ def get_distros(Map args) {
     }
 
     /// read distros from edition.yml otherwise.
-    inside_container() {
-        dir("${checkout_dir}") {
-            return cmd_output("""scripts/run-pipenv run \
-                  buildscripts/scripts/get_distros.py \
-                  --editions_file "${checkout_dir}/editions.yml" \
-                  use_cases \
-                  --edition "${edition}" \
-                  --use_case "${use_case}"
-            """).split().grep();
-        }
+    dir("${checkout_dir}") {
+        return cmd_output("""python3 \
+              buildscripts/scripts/get_distros.py \
+              --editions_file "${checkout_dir}/editions.yml" \
+              use_cases \
+              --edition "${edition}" \
+              --use_case "${use_case}"
+        """).split().grep();
     }
 }
 
 def get_editions() {
     /// read editions from edition.yml
-    inside_container() {
-        dir("${checkout_dir}") {
-            return cmd_output("""scripts/run-pipenv run \
-                  buildscripts/scripts/get_distros.py \
-                  --editions_file "${checkout_dir}/editions.yml" \
-                  editions
-            """).split().grep();
-        }
+    dir("${checkout_dir}") {
+        return cmd_output("""python3 \
+              buildscripts/scripts/get_distros.py \
+              --editions_file "${checkout_dir}/editions.yml" \
+              editions
+        """).split().grep();
     }
 }
 
 def get_internal_artifacts_pattern() {
-    inside_container() {
-        dir("${checkout_dir}") {
-            return sh(script: """scripts/run-pipenv run \
-                  buildscripts/scripts/get_distros.py \
-                  --editions_file "editions.yml" \
-                  internal_build_artifacts \
-                  --as-codename \
-                  --as-rsync-exclude-pattern;
-            """, returnStdout: true).trim();
-        }
+    dir("${checkout_dir}") {
+        return sh(script: """python3 \
+              buildscripts/scripts/get_distros.py \
+              --editions_file "editions.yml" \
+              internal_build_artifacts \
+              --as-codename \
+              --as-rsync-exclude-pattern;
+        """, returnStdout: true).trim();
     }
 }
